Refactor: numbers to .kv file, pie/numbers round
On a mobile device there is an inconsistency with buttons when
the app doesn't use Window.maximize(). The touch of the button
is not where the button appears. If you use Window.maximize(),
then the issue is solved, however then everything becomes oval.

So this is a change to keep the clock round while the window is
not square.  This required moving from pos_hint to pos based on
the minimum of height vs width. (Just assuming portrait mode at
this point. )